[CORE-8823] refactor, code review fixes
