+++ packages/inputs-minimal/LICENSE
