+++ packages/inputs-minimal/jsconfig.json
