+++ packages/inputs-minimal/package.json
