+++ packages/inputs-minimal/rollup.config.js
