+++ packages/inputs-minimal/src/constants.jsx
