+++ packages/inputs-minimal/src/index.jsx
@@ -1,0 +1,4 @@
+export { default as Radio } from './radio';
+export { default as Select } from './select';
+export { default as TextArea } from './textarea';
+export { default as TextField } from './textfield';
