+++ packages/inputs-minimal/src/select/index.jsx
@@ -1,0 +1,95 @@
+import PropTypes from 'prop-types';
+import clsx from 'clsx';
+import { SIZES, SIZES_KEYS } from '@/src/constants';
+import Meta from '../textfields_meta';
+import styles from './index.module.scss';
+
+const Select = ({
+  label,
+  error,
+  value = '',
+  options,
+  size = SIZES.medium,
+  onChange,
+  className,
+  disableBorder,
+  disabled,
+}) => {
+  const handleChange = (e) => {
+    const { selectedIndex } = e.target;
+    const selectedValue = options[selectedIndex].value || options[selectedIndex];
+    onChange(selectedValue, e);
+  };
+
+  return (
+    <div>
+      <div
+        className={clsx(
+          styles.inputWrapper,
+          styles[size],
+          {
+            [styles.disableBorder]: disableBorder,
+            [styles.withError]: error,
+            [styles.disabled]: disabled,
+          },
+        )}
+      >
+        <label
+          className={clsx(styles.label)}
+        >
+          <span
+            className={clsx(
+              styles.labelText,
+              styles[size],
+            )}
+          >
+            {label}
+          </span>
+
+          <select
+            className={clsx(styles.input, className)}
+            onChange={handleChange}
+            value={value}
+            disabled={disabled}
+          >
+            {options.map((option) => (
+              <option
+                key={option.value || option}
+                value={option.value || option}
+              >
+                {option.label || option}
+              </option>
+            ))}
+          </select>
+        </label>
+      </div>
+
+      <Meta
+        className={styles.meta}
+        error={error}
+      />
+    </div>
+  );
+};
+
+Select.propTypes = {
+  label: PropTypes.string,
+  error: PropTypes.string,
+  value: PropTypes.string,
+  options: PropTypes.oneOfType([
+    PropTypes.arrayOf(
+      PropTypes.shape({
+        label: PropTypes.string,
+        value: PropTypes.string,
+      }),
+    ),
+    PropTypes.array,
+  ]).isRequired,
+  onChange: PropTypes.func.isRequired,
+  size: PropTypes.oneOf(SIZES_KEYS),
+  className: PropTypes.string,
+  disableBorder: PropTypes.bool,
+  disabled: PropTypes.bool,
+};
+
+export default Select;
