+++ packages/inputs-minimal/src/select/index.module.scss
