+++ packages/inputs-minimal/src/select/index.stories.module.scss
