+++ packages/inputs-minimal/src/textarea/index.module.scss
