+++ packages/inputs-minimal/src/textarea/index.stories.module.scss
