+++ packages/inputs-minimal/src/textfield/index.stories.module.scss
@@ -1,0 +1,4 @@
+.container {
+  margin: 30vh 30vw 0;
+  position: relative;
+}
