--- packages/inputs-tiny/src/index.jsx
@@ -1,4 +1,0 @@
-/* eslint-disable import/no-unresolved,import/extensions */
-
-// glob gets magically expanded by rollup
-export * from './*/index.jsx';
